Split tests into separate attributes
Closes #25
