+++ test/Xunit.OpenCategories.UnitTests/DatabaseTestTests.cs
@@ -1,0 +1,16 @@
+﻿using FluentAssertions;
+
+namespace Xunit.OpenCategories.UnitTests;
+
+public class DatabaseTestTests
+{
+    [Fact]
+    [DatabaseTest]
+    public void DatabaseTest()
+    {
+        var testMethod = typeof(DatabaseTestTests).GetMethod(nameof(DatabaseTest));
+        testMethod.Should()
+            .BeDecoratedWith<FactAttribute>()
+            .And.BeDecoratedWith<DatabaseTestAttribute>();
+    }
+}
