+++ test/Xunit.OpenCategories.UnitTests/SnapshotTestTests.cs
@@ -1,0 +1,16 @@
+﻿using FluentAssertions;
+
+namespace Xunit.OpenCategories.UnitTests;
+
+public class SnapshotTestTests
+{
+    [Fact]
+    [SnapshotTest]
+    public void SnapshotTest()
+    {
+        var testMethod = typeof(SnapshotTestTests).GetMethod(nameof(SnapshotTest));
+        testMethod.Should()
+            .BeDecoratedWith<FactAttribute>()
+            .And.BeDecoratedWith<SnapshotTestAttribute>();
+    }
+}
